create typecheck.rs
