+++ node-graph/interpreted-executor/src/dynamic_executor.rs
@@ -4,8 +4,9 @@ use graph_craft::Type;
 use graph_craft::document::NodeId;
 use graph_craft::document::value::{TaggedValue, UpcastAsRefNode, UpcastNode};
 use graph_craft::graphene_compiler::Executor;
-use graph_craft::proto::{ConstructionArgs, GraphError, LocalFuture, NodeContainer, ProtoNetwork, ProtoNode, SharedNodeContainer, TypeErasedBox, TypingContext};
+use graph_craft::proto::{ConstructionArgs, GraphError, LocalFuture, NodeContainer, ProtoNetwork, ProtoNode, SharedNodeContainer, TypeErasedBox};
 use graph_craft::proto::{GraphErrorType, GraphErrors};
+use graph_craft::typecheck::TypingContext;
 use std::collections::{HashMap, HashSet};
 use std::error::Error;
 use std::sync::Arc;
